requirements.txt
torch==2.6.0
numpy==2.2.6 
scipy==1.15.3
h5py==3.15.1 
pyyaml==6.0.3
tqdm==4.67.3
wandb==0.24.2
opencv-python==4.13.0.90